fixed: I added protection for the findOne controller inside the wishes and also moved the config file to the config folder
