+++ src/auth/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { UsersService } from '../users/users.service';
-import config from '../config';
+import config from '../config/config';
 import * as dotenv from 'dotenv';
 dotenv.config();
 @Injectable()
